test: refactor libnetplan & keyfile parser testing a bit
